Rework training such that paralleltrainingsystem can become full pipeline ensemble
